# corex/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lab/analyze:
    post:
      operationId: analyzeLabProject
      tags: [generated-projects]
      summary: Analyze imported software evidence and propose a safe adaptation strategy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabAnalysisInput"
      responses:
        "200":
          description: Static observations and Router IA adaptation analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabAnalysisResult"
        "400":
          description: Invalid analysis request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Router IA could not analyze the evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /router/status:
    get:
      operationId: getRouterStatus
      tags: [router]
      security: [{ bearerAuth: [] }]
      summary: Get Router IA configuration and connection status
      responses:
        "200":
          description: Router IA configuration state; secret values are never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterConnectionStatus"
  /router/providers:
    get:
      operationId: listRouterProviders
      tags: [router]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User-owned providers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RouterProvider" }
    post:
      operationId: createRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouterProviderInput" }
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouterProvider" }
  /router/providers/{providerId}:
    parameters:
      - name: providerId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouterProviderUpdate" }
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouterProvider" }
    delete:
      operationId: deleteRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Provider deleted }
  /router/providers/{providerId}/test:
    parameters:
      - name: providerId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: testRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      deprecated: true
      responses:
        "410":
          description: Direct provider inference is disabled in CoreX
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /router/test:
    post:
      operationId: testRouterConnection
      tags: [router]
      security: [{ bearerAuth: [] }]
      summary: Check Router IA health without invoking a model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouterConnectionTestInput"
      responses:
        "200":
          description: Router IA health status; no prompt or provider inference is sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterConnectionStatus"
        "400":
          description: Invalid Router test request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /prisma/chat:
    post:
      operationId: sendPrismaChat
      tags: [prisma]
      security: [{ bearerAuth: [] }]
      summary: Send a conversational message through Router IA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrismaChatRequest"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrismaChatResponse"
        "400":
          description: Invalid chat request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Prisma module is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Per-user message limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Router IA request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    LabStaticEvidence:
      type: object
      properties:
        filePaths:
          type: array
          maxItems: 80
          items: { type: string, maxLength: 180 }
        dependencies:
          type: array
          maxItems: 80
          items: { type: string, maxLength: 160 }
        entryPoints:
          type: array
          maxItems: 40
          items: { type: string, maxLength: 180 }
        components:
          type: array
          maxItems: 60
          items: { type: string, maxLength: 180 }
        assets:
          type: array
          maxItems: 80
          items: { type: string, maxLength: 180 }
        strings:
          type: array
          maxItems: 100
          items: { type: string, maxLength: 220 }
        permissions:
          type: array
          maxItems: 60
          items: { type: string, maxLength: 180 }
        networkCalls:
          type: array
          maxItems: 60
          items: { type: string, maxLength: 220 }
        storage:
          type: array
          maxItems: 40
          items: { type: string, maxLength: 180 }
      required: [filePaths, dependencies, entryPoints, components, assets, strings, permissions, networkCalls, storage]
    LabAnalysisInput:
      type: object
      properties:
        sourceName:
          type: string
          minLength: 1
          maxLength: 180
        sourceKind:
          type: string
          enum: [apk, archive, code, document, image, website, binary]
        goal:
          type: string
          maxLength: 1600
        evidenceText:
          type: string
          minLength: 1
          maxLength: 30000
        evidence:
          $ref: "#/components/schemas/LabStaticEvidence"
        visuals:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/LabVisualEvidence"
      required: [sourceName, sourceKind, goal, evidenceText, evidence]
    LabVisualEvidence:
      type: object
      properties:
        sourceName:
          type: string
          minLength: 1
          maxLength: 180
        imageDataUrl:
          type: string
          maxLength: 1200000
          pattern: "^data:image/(jpeg|png|webp);base64,"
      required: [sourceName, imageDataUrl]
    LabAnalysisResult:
      type: object
      properties:
        summary:
          type: string
          maxLength: 1600
        architecture:
          type: string
          maxLength: 1600
        capabilities:
          type: array
          maxItems: 12
          items: { type: string, maxLength: 240 }
        observed:
          $ref: "#/components/schemas/LabStaticEvidence"
        inferred:
          type: array
          maxItems: 12
          items: { type: string, maxLength: 240 }
        reusableModules:
          type: array
          maxItems: 12
          items: { type: string, maxLength: 240 }
        adaptationPlan:
          type: array
          maxItems: 10
          items: { type: string, maxLength: 300 }
        risks:
          type: array
          maxItems: 10
          items: { type: string, maxLength: 240 }
        blockedOperations:
          type: array
          maxItems: 8
          items: { type: string, maxLength: 240 }
      required: [summary, architecture, capabilities, observed, inferred, reusableModules, adaptationPlan, risks, blockedOperations]
...
    RouterTaskType:
      type: string
      enum: [chat, coding, reasoning, summarization, vision, document, long_context, fast]
    RouterConnectionStatus:
      type: object
      properties:
        configured:
          type: boolean
        connected:
          type: boolean
        lastTestAt:
          type: ["string", "null"]
          format: date-time
        lastLatencyMs:
          type: ["integer", "null"]
          minimum: 0
        message:
          type: ["string", "null"]
          maxLength: 240
      required:
        - configured
        - connected
        - lastTestAt
        - lastLatencyMs
        - message
    RouterConnectionTestInput:
      type: object
      properties:
        task_type:
          type: string
          enum: [chat]
      required: [task_type]
    PrismaChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
          maxLength: 4000
      required: [role, content]
    PrismaChatRequest:
      type: object
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 60
          items:
            $ref: "#/components/schemas/PrismaChatMessage"
      required: [messages]
    PrismaChatResponse:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 8000
      required: [message]
    RouterProvider:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [openai, anthropic, gemini, openai-compatible, groq] }
        baseUrl: { type: ["string", "null"] }
        model: { type: string }
        apiKeyPreview: { type: string }
        isDefault: { type: boolean }
        capabilities: { type: array, minItems: 1, maxItems: 8, items: { $ref: "#/components/schemas/RouterTaskType" } }
        priority: { type: integer, minimum: 0, maximum: 100 }
        isActive: { type: boolean }
        status: { type: string }
        healthStatus: { type: string }
        lastTestAt: { type: ["string", "null"], format: date-time }
      required: [id, name, kind, baseUrl, model, apiKeyPreview, isDefault, capabilities, priority, isActive, status, healthStatus, lastTestAt]
    RouterProviderInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        kind: { type: string, enum: [openai, anthropic, gemini, openai-compatible, groq] }
        baseUrl: { type: string, maxLength: 500 }
        model: { type: string, minLength: 1, maxLength: 160 }
        apiKey: { type: string, minLength: 1, maxLength: 4096 }
        isDefault: { type: boolean }
        capabilities: { type: array, minItems: 1, maxItems: 8, items: { $ref: "#/components/schemas/RouterTaskType" } }
        priority: { type: integer, minimum: 0, maximum: 100 }
        isActive: { type: boolean }
      required: [name, kind, model, isDefault]
    RouterProviderUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        baseUrl: { type: string, maxLength: 500 }
        model: { type: string, minLength: 1, maxLength: 160 }
        apiKey: { type: string, minLength: 1, maxLength: 4096 }
        isDefault: { type: boolean }
        capabilities: { type: array, minItems: 1, maxItems: 8, items: { $ref: "#/components/schemas/RouterTaskType" } }
        priority: { type: integer, minimum: 0, maximum: 100 }
        isActive: { type: boolean }
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          maxLength: 240
      required: [error]