# corex/lib/api-zod/src/generated/types/routerProviderUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouterTaskType } from './routerTaskType';

export interface RouterProviderUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /** @maxLength 500 */
  baseUrl?: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  model?: string;
  /**
     * @minLength 1
     * @maxLength 4096
     */
  apiKey?: string;
  isDefault?: boolean;
  /**
     * @minItems 1
     * @maxItems 8
     */
  capabilities?: RouterTaskType[];
  /**
     * @minimum 0
     * @maximum 100
     */
  priority?: number;
  isActive?: boolean;
}