# corex/lib/api-zod/src/generated/types/routerTaskType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type RouterTaskType = typeof RouterTaskType[keyof typeof RouterTaskType];


export const RouterTaskType = {
  chat: 'chat',
  coding: 'coding',
  reasoning: 'reasoning',
  summarization: 'summarization',
  vision: 'vision',
  document: 'document',
  long_context: 'long_context',
  fast: 'fast',
} as const;